Change documentation for a url filter function
This threw us for a few minutes since it said URL could be either a boolean or an object with the key url and a function. Then later in the readme it says a shape of `{url: { filter: () => boolean } }.

This is also in the [webpack documentation](https://webpack.js.org/loaders/css-loader/#url) and I _think_ it's inaccurate.